search

+++ 0x0F-function_pointers/function_pointers.h
@@ -1,6 +1,6 @@
 #ifndef ALX_FILE
 #define ALX_FILE
 void print_name(char *name, void (*f)(char *));
-void array_iterator(int *array, size_t size, void (*action)(int));
+void array_iterator(int *array, int size, void (*action)(int));
 int int_index(int *array, int size, int (*cmp)(int));
 #endif
